Move extension files to new dir.
